Move some logic around
