+++ addon/utils/utils.js
@@ -51,8 +51,12 @@ export function isEmberObject(o) {
   return !!(o && o instanceof Ember.Object);
 }
 
-export default function isObject(o) {
+export function isObject(o) {
   return typeOf(o) === 'object' || typeOf(o) === 'instance';
+}
+
+export function isDescriptor(o) {
+  return o && typeof o === 'object' && o.isDescriptor;
 }
 
 export function isValidatable(value) {
