+++ addon/validators/base.js
@@ -5,6 +5,7 @@
 
 import Ember from 'ember';
 import Messages from 'ember-cp-validations/validators/messages';
+import Options from 'ember-cp-validations/-private/options';
 import { unwrapString, getValidatableValue, mergeOptions } from 'ember-cp-validations/utils/utils';
 
 const {
@@ -12,41 +13,8 @@ const {
   set,
   isNone,
   computed,
-  getOwner,
-  defineProperty
+  getOwner
 } = Ember;
-
-const Options = Ember.Object.extend({
-  model: null,
-  attribute: null,
-
-  // Private
-  __options__: null,
-
-  init() {
-    this._super(...arguments);
-
-    let options = this.get('__options__');
-
-    Object.keys(options).forEach((key) => {
-      let value = options[key];
-      if (value && typeof value === 'object' && value.isDescriptor) {
-        defineProperty(this, key, value);
-      } else {
-        set(this, key, value);
-      }
-    });
-  },
-
-  copy() {
-    let options = this.get('__options__');
-
-    return Ember.Object.create(Object.keys(options).reduce((obj, o) => {
-      obj[o] = get(this, o);
-      return obj;
-    }, {}));
-  }
-});
 
 /**
  * @class Base
@@ -147,24 +115,6 @@ const Base = Ember.Object.extend({
     // there is no need for it to be passed around
     this.value = builtOptions.value || this.value;
     delete builtOptions.value;
-
-    // let OptionsClass = Ember.Object.extend(builtOptions, {
-    //   model: computed(() => get(this, 'model')).readOnly(),
-    //   attribute: computed(() => get(this, 'attribute')).readOnly(),
-    //
-    //   copy(deep) {
-    //     if (deep) {
-    //       return OptionsClass.create();
-    //     }
-    //
-    //     return Ember.Object.create(Object.keys(builtOptions).reduce((obj, o) => {
-    //       obj[o] = get(this, o);
-    //       return obj;
-    //     }, {}));
-    //   }
-    // });
-    //
-    // return OptionsClass.create();
 
     return Options.create({
       model: get(this, 'model'),
